Commencement de la logique du cart + remaniement de certains models

+++ SERVER/routes/cartController.js
@@ -1,13 +1,27 @@
 const { Router } = require("express");
 const Cart = require("../models/cart");
+const CartProduct = require("../models/cartproduct");
 const checkAuth = require("../middlewares/checkAuth");
 const router = new Router();
 
 router.post("/", async (req, res, next) => {
   try {
-    const cart = await Cart.create(req.body);
+    const { id_user } = req.body;
+
+    if (!id_user) {
+      console.error("User ID is required");
+      return res.status(400).json({ error: "User ID is required" });
+    }
+
+    const expire_date = new Date(Date.now() + 15 * 60 * 1000);
+    const cart = await Cart.create({
+      id_user,
+      expire_date,
+    });
+
     res.status(201).json(cart);
   } catch (e) {
+    console.error("Error creating cart:", e);
     next(e);
   }
 });
@@ -61,17 +75,31 @@ router.delete("/:id", async (req, res, next) => {
   }
 });
 
-router.put("/:id", async (req, res, next) => {
+router.put("/", async (req, res, next) => {
   try {
+    const { id_user } = req.body;
+    console.log(`Received request to update cart for user ID: ${id_user}`);
+
+    if (!id_user) {
+      console.error("User ID is required");
+      return res.status(400).json({ error: "User ID is required" });
+    }
+
+    // Supprimez les paniers existants pour cet utilisateur
     const nbDeleted = await Cart.destroy({
       where: {
-        id: parseInt(req.params.id),
+        id_user: id_user,
       },
     });
+
+    // Créez un nouveau panier pour cet utilisateur
+    const expire_date = new Date(Date.now() + 15 * 60 * 1000);
     const cart = await Cart.create({
-      ...req.body,
-      id: parseInt(req.params.id),
+      id_user,
+      expire_date,
     });
+
+    // Répondez avec le nouveau panier
     res.status(nbDeleted ? 200 : 201).json(cart);
   } catch (e) {
     next(e);
